+++ lib/genesis_ruby/api/constants/non_financial/kyc/profile_current_statuses.rb
@@ -1,0 +1,25 @@
+module GenesisRuby
+  module Api
+    module Constants
+      module NonFinancial
+        module Kyc
+          # Profile current statuses
+          class ProfileCurrentStatuses
+
+            extend Mixins::Constants::Common
+
+            # Undefined
+            UNDEFINED = 0
+            # Review
+            REVIEW    = 1
+            # Denied
+            DENIED    = 2
+            # Approved
+            APPROVED  = 3
+
+          end
+        end
+      end
+    end
+  end
+end