+++ lib/genesis_ruby/api/requests/financial/preauthorizations/partial_reversal.rb
@@ -1,0 +1,24 @@
+module GenesisRuby
+  module Api
+    module Requests
+      module Financial
+        module Preauthorizations
+          # Partial Reversal transaction type
+          class PartialReversal < Base::Reference
+
+            # Partial Reversal transaction type
+            def transaction_type
+              Constants::Transactions::PARTIAL_REVERSAL
+            end
+
+            # Returns the transaction structure
+            def reference_transaction_structure
+              {}
+            end
+
+          end
+        end
+      end
+    end
+  end
+end