+++ spec/genesis_ruby/api/constants/non_financial/kyc/document_supported_types_spec.rb
@@ -1,0 +1,7 @@
+require 'genesis_ruby/api/constants/non_financial/kyc/document_supported_types'
+
+RSpec.describe GenesisRuby::Api::Constants::NonFinancial::Kyc::DocumentSupportedTypes do
+
+  include_examples 'common constants examples', described_class::CREDIT_OR_DEBIT_CARD
+
+end