+++ spec/genesis_ruby/stubs/api/requests/non_financial/kyc/document_attributes_stub.rb
@@ -1,0 +1,21 @@
+module GenesisSpec
+  module Stubs
+    module Api
+      module Requests
+        module NonFinancial
+          module Kyc
+            # Stub class for DocumentAttributesStub mixin
+            class DocumentAttributesStub
+
+              include GenesisRuby::Api::Mixins::Requests::RestrictedSetter
+              include GenesisRuby::Api::Mixins::Requests::NonFinancial::Kyc::DocumentAttributes
+
+              public :document_attributes_structure
+
+            end
+          end
+        end
+      end
+    end
+  end
+end